% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVector.R
\name{colorVector}
\alias{colorVector}
\title{Creates a vector of colors of the same length as the data}
\usage{
colorVector(x, order = unique(x))
}
\arguments{
\item{x}{A vector, in most cases of identities of individuals or clusters ect.}

\item{order}{The order, folowing a rainbow distribution, that the colors should be in in the output vector. Defaults to the order that the unique values in x occurs.}
}
\value{
A vector, the same length as x with each unique value substitutet with a color.
}
\description{
This function takes a vector x and a shorter ordering vector with all the unique values of the x vector in the specific order that the colors should be in and returns a vector of RGB colors the same length as the initial x vector.
}
\examples{
#Generate a dataframe with bimodally distributed data and a few separate subsamplings
x <- generateFlowCytometryData(samplings=5, observations=2000)

#Scale the data (not actually necessary in this artificial 
#example due to the nature of the generated data)
x_scaled <- quantileScale(x=x[2:ncol(x)])

#Run Barnes Hut tSNE on this. 
library(Rtsne.multicore)
xSNE <- Rtsne.multicore(x_scaled, pca=FALSE)

#Now use our function
xColors <- colorVector(x[,1])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Plot all ids together and use rainbowColors
depecheDensity(xYData=as.data.frame(xSNE$Y), commonName="All_samplings", 
color=xColors, createDirectory=FALSE)

}
