% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dClust.R
\name{dClust}
\alias{dClust}
\title{Function to run penalized K means}
\usage{
dClust(inDataFrameScaled, dOptObject, ids, sampleSize, penalty, withOrigoClust,
  initCenters = 30)
}
\arguments{
\item{inDataFrameScaled}{A dataframe with the data that will be used to create the clustering. The data in this dataframe should be scaled in a proper way. Empirically, many datasets seem to be clustered in a meaningful way if they are scaled with the dScale function.}

\item{dOptObject}{This object contains information about optimal sample size, penalty offset, solution with or without a cluster in origo, and the number of initial cluster centers that were used to find this optimal information.}

\item{ids}{A vector of the same length as rows in the inDataFrameScaled. It is used to generate the final analysis, where a table of the percentage of observations for each individual and each cluster is created.}

\item{sampleSize}{By default inherited from dOptObject. Number of observations that shoult be included in the initial clustering step. Three possible values. Either inherited, "All" or a user-specified number. Defaults to inheriting from dClustObject. If a dClustObject is not substituted, all rows in inDataFrameScaled are added by default. If another number, a sample is created from inDataFrameScaled. This is extra useful when clustering very large datasets. Replacement is set to TRUE.}

\item{penalty}{By default inherited from dOptObject. The parameter that controls the level of penalization.}

\item{withOrigoClust}{By default inherited from dOptObject. This parameter controls if the generated result should contain a cluster in origo or not.}

\item{initCenters}{By default inherited from dOptObject. Number of starting points for clusters. This essentially means that it is the highest possible number of clusters that can be defined. The higher the number, the greater the precision, but the computing time is also increased with the number of starting points. Default is 30.}
}
\value{
A list with three components:
\describe{
    \item{clusterVector}{A vector with the same length as number of rows in the inDataFrameScaled, where the cluster identity of each observation is noted.}
    \item{clusterCenters}{A matrix containing information about where the centers are in all the variables that contributed to creating the cluster with the given penalty term.}
    \item{clusterPercentagesForAllIds}{A matrix showing the percentage of observations for each id in each cluster.}
}
}
\description{
This function is the core user function of the Depeche package. It clusters the data with a penalized version of K-means.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateBimodalData(samplings=2)

#Scale this datamframe
x_scaled <- dScale(x[,2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the dOptPenalty function to get good starting points
x_optim <- dOpt(x_scaled)

#Then run the actual function
x_dClust <- dClust(x_scaled, dOptObject=x_optim, ids=x[,1])

#And finally look at your great result
str(x_dClust)
}
\seealso{
\code{\link{dAllocate}}, \code{\link{dOpt}}, \code{\link{dOptAndClust}}
}
