% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depecheColor.R
\name{depecheColor}
\alias{depecheColor}
\title{Display third variable as color on 2D plot}
\usage{
depecheColor(colorData, names = "default", xYData, title = FALSE,
  dotSize = 20000/nrow(colorData), drawColorPalette = FALSE,
  createDirectory = TRUE,
  directoryName = "Variables displayed as color on SNE field")
}
\arguments{
\item{colorData}{A vector or a dataframe of numeric observations that will be displayed as color on the plot.}

\item{names}{The name(s) for the plots. The default alternative, "default" returns the column names of the colorData object in the case this is a dataframe and otherwise returns the somewhat generic name "testVariable". It can be substitutet with a string (in the case colorData is a vector) or vector of strings, as long as it has the same length as the number of columns in colorData.}

\item{xYData}{These variables create the field on which the colorData will be displayed. It needs to be a dataframe with two columns and the same number of rows as the colorData object.}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved a jpeg, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}

\item{drawColorPalette}{If a separate plot with the color palette used for the plots should be printed and saved.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}
}
\value{
Plots showing the colorData displayed as color on the field created by xYData.
}
\description{
Function to overlay one variable for a set of observations on a field created by two other variables known for the same observations. The plot is constructed primarily for displaying variables on 2D-stochastic neighbour embedding fields, but can be used for any sets of (two or) three variables known for the same observations. As the number of datapoints is often very high, the files would, if saved as pdf of another vector based file type become extremely big. For this reason, the plots are saved as jpeg and no axes or anything alike are added, to simplify usage in publications.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateFlowCytometryData()

#Scale the data (not actually necessary in this artificial example due to the nature of the generated data)
x_scaled <- quantileScale(x[2:ncol(x)])

#Run Barnes Hut tSNE on this. NB! This takes quite a while (a minute or so) as the algorithm is slow.
library(Rtsne)
xSNE <- Rtsne(x_scaled, pca=FALSE)

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the function for all the variables
depecheColor(colorData=x_scaled, xYData=as.data.frame(xSNE$Y), drawColorPalette=TRUE)

}
\seealso{
\code{\link{depecheDensity}}, \code{\link{depecheResidual}}, \code{\link{depecheWilcox}}
}
