% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depecheDensity.R
\name{depecheDensity}
\alias{depecheDensity}
\title{Display density on 2D plot}
\usage{
depecheDensity(xYData, scalingControl, plotEachIdSeparately = FALSE,
  idsVector, commonName, color = c("blue", "rainbowCols", "idSpecific",
  "multipleColors"), densContour, bandColor = "black",
  dotSize = 400/sqrt(nrow(xYData)), title = FALSE, createDirectory = TRUE,
  directoryName = paste("Density plots for ", commonName, "s", sep = ""))
}
\arguments{
\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation.}

\item{scalingControl}{A dataframe with two columns. This argument is only necessary if the xYData should be scaled to another range than its internal maximum and minimum values.}

\item{plotEachIdSeparately}{If the xYData is made up of data from multiple, individual subjects, these can be plotted separately. Defaults to FALSE.}

\item{idsVector}{If "plotEachIdSeparately" is TRUE, then this argument is needed to provide information about which rows that belong to which id.}

\item{commonName}{A name that is common to all density plots created. It can be the groups name, e.g. "Malaria patients" or "Clusters". If only one plot is created, the name is still taken from here.}

\item{color}{This gives the specific color of the plot. It has four potential return values. Either it is a specific color ("red" or "#FF0000"), it is "rainbowCols", which creates a specific set of colors similar to the rainbow. The third alternative is "idSpecific", which will hten give each plot a different color depending on their id. The fourth alternative is that a vector of colors (e.g. either "red" or "#FF0000") with the same length as the number of rows in xYData is provided. Here, a separate density estimate will be performed for each color subcompartment of the xYData and these will be plotted together.}

\item{densContour}{An object to create the density contours for the plot. If not present, it will be generated with the xYData. Useful when only a subfraction of a dataset is plotted, and a superimposition of the distribution of the whole dataset is of interest.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}
}
\value{
Plots showing the densities of the specific xYData (subset) displayed as color on the field created by the same xYData (subset).
}
\description{
Function to show density for a set of observations on a field created by two variables. The plot is constructed primarily for displaying density of 2D-stochastic neighbour embedding fields, but can be used for any sets of two known for the same observations. As the number of datapoints is often very high, the files would, if saved as pdf of another vector based file type become big. For this reason, the plots are saved as jpeg and no axes or anything alike are added, to simplify usage in publications.
}
\examples{
#Generate a dataframe with bimodally distributed data and a few separate subsamplings
x <- generateFlowCytometryData(samplings=5, observations=2000)

#Scale the data (not actually necessary in this artificial example due to the nature of the generated data)
x_scaled <- quantileScale(x[2:ncol(x)])

#Run Barnes Hut tSNE on this. NB! This takes quite a while (a minute or so) as the algorithm is slow.
library(Rtsne)
xSNE <- Rtsne(x_scaled, pca=FALSE)

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Plot all ids together and use rainbowColors
sneDensityPlot(xYData=as.data.frame(xSNE$Y), commonName="All_samplings", color="rainbowCols", createDirectory=FALSE)

#Now plot each id separately
sneDensityPlot(xYData=as.data.frame(xSNE$Y), plotEachIdSeparately=TRUE, idsVector=x[,1], commonName="sampling", color="idSpecific")

}
\seealso{
\code{\link{depecheColor}}, \code{\link{depecheResidual}}, \code{\link{depecheWilcox}}
}
