% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depecheResidual.R
\name{depecheResidual}
\alias{depecheResidual}
\title{Showing the residuals when subtracting the values from one group from another on a sne plot}
\usage{
depecheResidual(groupClusterData, xYData, densContour,
  name = "depecheResidual", groupName1 = "Group 1",
  groupName2 = "Group 2", title = FALSE, maxAbsPlottingValues,
  bandColor = "black", createDirectory = FALSE,
  directoryName = "depecheResidual", dotSize = 400/sqrt(nrow(xYData)))
}
\arguments{
\item{groupClusterData}{This dataframe should contain at least two columns  with information about the cluster (named "cluster") and the group identity (named "group") for each observation.}

\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation. It needs to have the same number of rows as groupClusterData.}

\item{densContour}{An object to create the density contours for the plot. If not present, it will be generated with the xYData. Useful when only a subfraction of a dataset is plotted, and a superimposition of the distribution of the whole dataset is of interest.}

\item{name}{The main name for the graph and the analysis.}

\item{groupName1}{The name for the first group}

\item{groupName2}{The name for the second group}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{maxAbsPlottingValues}{If multiple plots should be compared, it might be useful to define a similar color scale for all plots, so that the same color always means the same value. Such a value can be added here. It defaults to the maximum Wilcoxon statistic that is generated in the analysis.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}
}
\value{
A sne based plot showing which events that belong to a cluster dominated by the first or the second group.
}
\description{
This function is used to compare groups of individuals from whom comparable cytometry or other complex data has been generated where the number of individuals does not allow any statistical comparisons.
}
\examples{
#Generate a dataframe with bimodally distributed data and 2 subsamplings.
x <- generateFlowCytometryData(samplings=2, ncols=7)

#Scale the data (not actually necessary in this artificial 
#example due to the nature of the generated data)
x_scaled <- quantileScale(x=x[2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Create the optimized number of clusters for this dataset
x_optim <- pKMOptim(x_scaled, iterations=10, bootstrapObservations=1000)
x_pKM <- pKMRun(x_scaled, regVec=x_optim[[1]][["optimalRegularizationValue"]], 
withOrWithoutZeroClust=x_optim[[1]][["withOrWithoutZeroClust"]], iterations=1, ids=x[,1])

#Run Barnes Hut tSNE on this. 
library(Rtsne.multicore)
xSNE <- Rtsne.multicore(x_scaled, pca=FALSE)

#Create the groupClusterData object
groupClusterData <- as.data.frame(cbind(x[,1], x_pKM$clusterVector))
colnames(groupClusterData) <- c("group", "cluster")

#And finally run the function
depecheResidual(groupClusterData=groupClusterData, xYData=as.data.frame(xSNE$Y))
}
\seealso{
\code{\link{depecheColor}}, \code{\link{depecheDensity}}, \code{\link{depecheWilcox}}
}
