% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depecheWilcox.R
\name{depecheWilcox}
\alias{depecheWilcox}
\title{Wilcoxon rank-sum or signed rank test comparison of subject groups in pKMRun result}
\usage{
depecheWilcox(idGroupClusterData, paired = FALSE,
  multipleCorrMethod = "hochberg", xYData, densContour,
  name = "depecheWilcox", groupName1 = "Group 1", groupName2 = "Group 2",
  title = FALSE, maxAbsPlottingValues, bandColor = "black",
  createDirectory = FALSE, directoryName = "depecheWilcox",
  dotsize = 400/sqrt(nrow(xYData)))
}
\arguments{
\item{idGroupClusterData}{This dataframe should contain three columns with information about the cluster identity (named "cluster"), the id (named "ids")and the group (named "group")identity for each observation.}

\item{paired}{If the data is paired, so that Wilcoxon signed rank test instead of Wilcoxon rank-sum test/Mann_Whitney test can be used. Defaults to false, i.e. no assumption of pairing is made and Wilcoxon rank sum-test.}

\item{multipleCorrMethod}{Which method that should be used for adjustment ofmultiple comparisons. Defaults to Benjamini-Hochberg, but all other methods available in \code{\link{p.adjust}} can be used.}

\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation. It needs to have the same number of rows as idGroupClusterData.}

\item{densContour}{An object to create the density contours for the plot. If not present, it will be generated with the xYData. Useful when only a subfraction of a dataset is plotted, and a superimposition of the distribution of the whole dataset is of interest.}

\item{name}{The main name for the graph and the analysis.}

\item{groupName1}{The name for the first group}

\item{groupName2}{The name for the second group}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{maxAbsPlottingValues}{If multiple plots should be compared, it might be useful to define a similar color scale for all plots, so that the same color always means the same statistical value. Such a value can be added here. It defaults to the maximum Wilcoxon statistic that is generated in the analysis.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}
}
\value{
This function always returns a dataframe showing the Wilcoxon statistic and the p-value for each cluster, with an included adjustment for multiple comparisons (see above). It also returns a sne based plot showing which events that belong to a cluster dominated by the first or the second group.
}
\description{
This function is used to compare groups of individuals from whom comparable cytometry or other complex data has been generated.
}
\examples{
#Generate a dataframe with bimodally distributed data and 20 subsamplings.
xindividuals <- generateFlowCytometryData(samplings=40, ncols=7, observations=500)

#Now add three columns that will separate the first ten from the second ten individuals and merge the datasets
xgroups <- generateFlowCytometryData(samplings=2, ncols=3, observations=10000)
colnames(xgroups)[2:4] <- c("X8", "X9", "X10")
x <- cbind(xindividuals[,1], xgroups[,1], xindividuals[,2:ncol(xindividuals)], xgroups[,2:ncol(xgroups)])
colnames(x)[1:2] <- c("ids", "group")

#Scale the data (not actually necessary in this artificial example due to the nature of the generated data)
x_scaled <- quantileScale(x[3:ncol(x)])

#Create the optimized number of clusters for this dataset
x_optim <- pKMOptim(x_scaled, iterations=50, bootstrapObservations=1000)
x_pKM <- pKMRun(x_scaled, regVec=x_optim[[1]][["optimalRegularizationValue"]], withOrWithoutZeroClust=x_optim[[1]][["withOrWithoutZeroClust"]], iterations=1, ids=x[,1])

#Run Barnes Hut tSNE on this dataset. XXX NB! This takes quite a while (two minutes or so) as the algorithm is slow.
library(Rtsne)
xSNE <- Rtsne(x_scaled, pca=FALSE)

#Create the idGroupClusterData object
idGroupClusterData <- cbind(x[,1:2], x_pKM$clusterVector)
colnames(idGroupClusterData)[3] <- "cluster"

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#And finally run the function
depecheWilcox(idGroupClusterData=idGroupClusterData, xYData=as.data.frame(xSNE$Y))
}
\seealso{
\code{\link{depecheColor}}, \code{\link{depecheDensity}}, \code{\link{depecheResidual}}
}
