% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOpt.R
\name{dOpt}
\alias{dOpt}
\title{Find the smallest optimal sampe size and the best penalty offset for a certain dataset}
\usage{
dOpt(inDataFrameScaled, initSampleSize = 2000, sampleSizePowerIncrement = 1,
  maxSampleSize = 1e+05, initCenters = 30, maxIter = 100,
  minImprovement = 0.01, penaltyOptOnly = FALSE, penalties = c(0, 2, 4, 8,
  16, 32, 64, 128))
}
\arguments{
\item{inDataFrameScaled}{A dataframe with the data that will be used to create the clustering. The data in this dataframe should be scaled in a proper way. Empirically, many datasets seem to be clustered in a meaningful way if they are scaled with the dScale function.}

\item{initSampleSize}{The initial sample size. If penaltyOptOnly==TRUE, this value will define the sample size for the penalty optimization. Defaults to 2000.}

\item{sampleSizePowerIncrement}{With what value the sample size will increase. The value is taken to the power of two, i e, the standard value of 1 translates to a start value of x*2^0, and the following value of 2^(0+1), and so on.}

\item{maxSampleSize}{At what value the algorithm stops. Defaults to 100 000.}

\item{initCenters}{Number of starting points for clusters. The higher the number, the greater the precision of the clustering, but the computing time is also increased with the number of starting points. Default is 30.}

\item{maxIter}{The maximal number of iterations that are performed to reach the minimal improvement.}

\item{minImprovement}{This is connected to the evaluation of the performance of the algorithm. The more iterations that are run, or the larger the samples are, the smaller will the improvement be, and this sets the threshold when the iterations stop.}

\item{penaltyOptOnly}{If this is set to true, no optimization on the sample size is performed. This might be useful in situations when the whole dataset will be used for clustering, such as when the dataset is small, or when the number of variables is very large (>100), so that increasing the sample size too far might lead to computational overload. If this is set to true, the sample size will be the value specified in the sampleSizeIncrement argument.}

\item{penalties}{These values are the ones that are evaluated and the ones that decide the penalization. The number of suggested default values are empirically defined and might not be optimal for a specific dataset, but the algorithm will warn if the most optimal values are on the borders of the range. Note that when this offset is 0, there is no penalization, which means that the algorithm runs normal K-means clustering.}
}
\value{
If penaltyOptOnly is set to FALSE, which is default, one graph showing the stability of the models with different sample sizes, one showing the distance with different penalties before the optimization of the sample size, and a third showing the same, but after the sample size has been optimized. In addition, a list is returned with the following content:
\describe{
    \item{sampleSizeOpt}{A dataframe, in which each row represents one sample size, and in which the last row is thus the chosen, optimal sample size. It has the following columns:
    \describe{
              \item{SampleSize}{This column shows the sample size of each boot strap subsampling in the optimization procedure.}
              \item{Lowest distance}{This vector shows the optimal stability, expressed as the lowest distance between the bootstrap subsampling runs at each of the boot strap subsamling sizes.}
              \item{Improvement}{Here, the improvement, expressed as a fraction between 0 and 1 is shown. When the improvement is less than minImprovement, the algorithm automatically stops.}
             } 
    }
    \item{penaltyOpt.df}{A dataframe with one row with all the information about which settings that were used to generate the optimal clustering with the optimal sample size. The "withOrigoClust" information tells the user if the solution with or without a cluster in origo gives the most optimal solution. If yes, this origo population is generally small and could be viewed as not fitting in the model.}
    \item{meanOptimDf}{A dataframe with the information about the results with all tested penalty values}
}

If instead penaltyOptOnly is set to TRUE, one graph showing the distance with different penalties as well as the two last items in the list described above is returned.
}
\description{
This function is used before the dClust function to identify the smallest sample size that gives rise to the most stable clustering. It is primarily used when datasets are very large (i.e. >100 000 observations), where there is a computational gain in clustering based on a subset of the cells and then assigning all other cells to the correct cluster center. It can however be generally useful when it is not entirely clear which sample size that is needed to perform the optimization of the penalty term, something that this function also works as a wrapper for.
}
\examples{
#Generate a dataframe with bimodally distributed data with a million rows.
x <- generateBimodalData(observations=1000000)

#Scale this datamframe
x_scaled <- dScale(x[,2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the function to identify at what sample size the cluster stability plateaus
x_optim <- dOpt(x_scaled)
}
\seealso{
\code{\link{dClust}}, \code{\link{dAllocate}}, \code{\link{dOptAndClust}}
}
