% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOptSampleSize.R
\name{dOptSampleSize}
\alias{dOptSampleSize}
\title{Fint the smallest optimal sampe size}
\usage{
dOptSampleSize(inDataFrameScaled, sampleSize = 1000 * c(2^0, 2^1, 2^2, 2^3,
  2^4, 2^5, 2^6, 2^7, 2^8, 2^9, 2^10), initCenters = 30, iterations = 50,
  minimalImprovement = 0.01, penaltyOptSampleSize = 16000,
  penaltyOffset = c(0, 2, 4, 8, 16, 32, 64, 128), makeOptimGraph = TRUE)
}
\arguments{
\item{inDataFrameScaled}{A dataframe with the data that will be used to create the clustering. The data in this dataframe should be scaled in a proper way. Empirically, many datasets seem to be clustered in a meaningful way if they are scaled with the quantileScale function.}

\item{sampleSize}{The number of observations that are included in each bootstrap subsampling of the data. NB! The algorithm uses resampling, so the same event can be used twice. This is the central argument to this function, that it optimizes over.}

\item{initCenters}{Number of starting points for clusters. This essentially means that it is the highest possible number of clusters that can be defined. The higher the number, the greater the precision, but the computing time is also increased with the number of starting points. Default is 30}

\item{iterations}{As it sounds, the number of bootstrap reiterations that are performed.}

\item{minimalImprovement}{Expressed as a fraction between 0 and 1, this value tells how much better a doubling of the sample size needs to be for the algorithm to continue to another step. Default is 0.001, corresponding to an improvement of one percent.}

\item{penaltyOptSampleSize}{The size of the initial samle that will be used to define the penalty offset that will be used in all subsequent optimizations of the sample size.}

\item{penaltyOffset}{These values are the ones that are evaluated and the ones that decide the penalization. The number of suggested default values are empirically defined and might not be optimal for a specific dataset, but the algorithm will warn if the most optimal values are on the borders of the range. Note that when this offset is 0, there is no penalization, which means that the algorithm runs normal K-means clustering.}

\item{makeOptimGraph}{If a graph should be created showing the distance between bootstraps under different penalties during the initial pre-optimization.}
}
\value{
A graph showing the stability of the models with different sample sizes. Optionally also the optimization graphs for each step, see makeOptimGraphs above. In addition, a data frame is returned. This one has three columns:
\describe{
    \item{SampleSize}{This column shows the sample size of each boot strap subsampling in the optimization procedure.}
    \item{Lowest distance}{This vector shows the optimal stability, expressed as the lowest distance between the bootstrap subsampling runs at each of the boot strap subsamling sizes.}
    \item{Improvement}{Here, the improvement, expressed as a fraction between 0 and 1 is shown. When the improvement is less than 1 percent, the algorithm automatically stops.}
}
}
\description{
This function is used before the dClust function to identify the smallest sample size that gives rise to the most stable clustering. It is primarily used when datasets are very large (i.e. >100 000 observations), where there is a computational gain in clustering based on a subset of the cells and then assigning all other cells to the correct cluster center.
}
\examples{
#Generate a dataframe with bimodally distributed data with a million rows.
x <- generateFlowCytometryData(observations=1000000)

#Scale this datamframe
x_scaled <- quantileScale(x[,2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the function to identify at what sample size the cluster stability plateaus
x_optim <- dOptSampleSize(x_scaled)
}
\seealso{
\code{\link{dClust}}, \code{\link{pClustPredict}}
}
