% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dViolins.R
\name{dViolins}
\alias{dViolins}
\title{Create violin plots for all non-penalized variable for all clusters}
\usage{
dViolins(depecheObject, order = unique(clusterVector),
  colorScale = "viridis", plotAll = FALSE)
}
\arguments{
\item{depecheObject}{A list object generated with the depeche function, containing a cluster vector and a cluster centers matrix and the scaled data used for clustering.}

\item{order}{The order that the unique features of the cluster vector should appear in. For harmonization with colorVector and all subsequent functions.}

\item{colorScale}{The color scale. Inherited from the viridis, gplots and grDevices packages (and the package-specific "dark_rainbow"). Seven possible scales are pre-made: inferno, magma, plasma, viridis, rich_colors, rainbow and dark_rainbow. User specified vectors of colors (e.g. c("#FF0033", "#03AF49")) are also accepted.}

\item{plotAll}{If all parameters, including the non-contributing, should be plotted for each cluster. Defaults to FALSE.}
}
\value{
One graph is created for each non-penalized variable in each non-penalized cluster, which often means that the function creates a vast number of graphs. The graphs are sorted into subfolders for each cluster.
}
\description{
Here, violin plots of a specific cluster and the total population are created for each variable that has not been penalized away in the penalized K-means analysis. As al such plots are generated for each cluster, this function creates a great number of plots in most instances.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateBimodalData(samplings=2, dataCols=8)

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Optimize and run the clustering function.
xDepecheObject <- depeche(x[2:ncol(x)])

#Create the plots of the variables that contribute to creating each cluster
dViolins(xDepecheObject)

#Now create plots of all clusters, regardless of if they contributed or not
dViolins(xDepecheObject, plotAll=TRUE)
}
\seealso{
\code{\link{dDensityPlot}}, \code{\link{dColorPlot}}, \code{\link{colorVector}}
}
