% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBimodalData.R
\name{generateBimodalData}
\alias{generateBimodalData}
\title{Function to generate synthetic flow cytometry-like data}
\usage{
generateBimodalData(samplings = 1, dataCols = 5,
  observations = 10000)
}
\arguments{
\item{samplings}{Number of individual probability samplings with identical standard deviation samplings, ie a simple simulated parallel to number of subjects.}

\item{dataCols}{Number of columns in the generated dataframe. Default is 5.}

\item{observations}{Number of rows per donor in the generated dataframe. Default is 10000.}
}
\value{
A dataframe with the specified numbers of columns and rows.
}
\description{
This function creates a dataframe with any chose size with bimodally distributed data in each column. The probability of the sizes of the respective populations in the bimodal distributions is stochastic.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateBimodalData()

#Plot the first two variables in this dataframe
plot(x[,2], x[,3])

}
