% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxPeakScale.R
\name{maxPeakScale}
\alias{maxPeakScale}
\title{Scaling and centering around the peak of the data}
\usage{
maxPeakScale(x, control, lowQuantile = 0.001, highQuantile = 0.999)
}
\arguments{
\item{x}{A numeric/integer vector or dataframe}

\item{control}{A numeric/integer vector or dataframe of values that could be used to define the range. If no control data is present, the function defaults to using the indata as control data.}

\item{lowQuantile}{The lower border below which the values are treated as outliers and will be outside of the defined scaling range (0-1*multiplicationFactor).}

\item{highQuantile}{The higher border above which the values are treated as outliers and will be outside of the defined scaling range (0-1*multiplicationFactor).}
}
\value{
A vector or dataframe with the same size but where all values in the vector or column of the dataframe have been internally scaled.
}
\description{
This function scales the data to the standard deviation within a default or user-defined quantile range. It the centers the data aroun the highest peak within it.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateFlowCytometryData()

#The maximum and minimum values are
summary(x)

#Run the function
y <- maxPeakScale(x[2:ncol(x)], lowQuantile=0.01, highQuantile=0.99)

#And the data has been rescaled to the standard deviation of range between the first and the 99:th percentile 
#of the data in the original dataframe columns. The data has also been centered around the peak of each column.
summary(y)

}
\seealso{
\code{\link{minMaxScale}} \code{\link{quantileScale}}
}
