% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pKMOptim.R
\name{pKMOptim}
\alias{pKMOptim}
\title{Function to find the optimal penalty value for penalized K means}
\usage{
pKMOptim(inDataFrameScaled, kVec = 30, iterations = 50,
  bootstrapObservations = 10000, regVecOffset = c(0, 5, 10, 15, 20, 25, 30,
  35, 40, 45, 50))
}
\arguments{
\item{inDataFrameScaled}{A dataframe with the data that will be used to create the clustering. The data in this dataframe should be scaled in a proper way. Empirically, many datasets seem to be clustered in a meaningful way if they are scaled with the quantileScale function.}

\item{kVec}{Number of starting points for clusters. This essentially means that it is the highest possible number of clusters that can be defined. The higher the number, the greater the precision, but the computing time is also increased with the number of starting points. Default is 30}

\item{iterations}{As it sounds, the number of bootstrap reiterations that are performed.}

\item{bootstrapObservations}{The number of observations that are included in each bootstrap subsampling of the data. NB! The algorithm uses resampling, so the same event can be used twice.}

\item{regVecOffset}{These values are the ones that are evaluated and the ones that decide the penalization. The number of suggested default values are empirically defined and might not be optimal for a specific dataset, but the algorithm will warn if the most optimal values are on the borders of the range. Note that when this offset is 0, there is no penalization, which means that the algorithm runs normal K-means clustering.}
}
\value{
A graph showing the performance of the algorithm under the different regVec values and a list with two components:
\describe{
    \item{regOpt.df}{A dataframe with one row with all the information about which settings that were used to generate the optimal clustering. The "withOrWithoutZeroClust" information tells the user if the solution with or without a cluster in origo gives the most optimal solution. Generally it is the clustering without an origo cluster ("stabWOZero) that is optimal.}
    \item{meanOptimDf}{A dataframe with the information about the results with all tested regVec values}
}
}
\description{
This function is used before the pKMRun function to identify the optimal regVec value for the specific dataset. This value decides the penalization and consequently also the number of clusters that are identified.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateFlowCytometryData()

#Scale this datamframe
x_scaled <- quantileScale(x[,2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the function
x_optim <- pKMOptim(x_scaled, bootstrapObservations=1000)
}
\seealso{
\code{\link{pKMRun}}, \code{\link{pKMPredict}}
}
