% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pKMRun.R
\name{pKMRun}
\alias{pKMRun}
\title{Function to run penalized K means}
\usage{
pKMRun(inDataFrameScaled, regVec, withOrWithoutZeroClust, kVec = 30,
  iterations = 10, ids)
}
\arguments{
\item{inDataFrameScaled}{A dataframe with the data that will be used to create the clustering. The data in this dataframe should be scaled in a proper way. Empirically, many datasets seem to be clustered in a meaningful way if they are scaled with the quantileScale function.}

\item{regVec}{The parameter that controls the level of penalization. Preferrably, it should be inherited from a pKMOptim run, as the algorithm will then generate the most stable result.}

\item{withOrWithoutZeroClust}{This parameter controls if the generated result should contain a cluster in origo or not. This information is given by pKMOptim, again.}

\item{kVec}{Number of starting points for clusters. This essentially means that it is the highest possible number of clusters that can be defined. The higher the number, the greater the precision, but the computing time is also increased with the number of starting points. Default is 30.}

\item{iterations}{As it sounds, this controls how many iterations that are performed, among which the most stable is chosen. If pKMOptim has been performed before, this number should not need to be extensive. Default is 10.}

\item{ids}{A vector of the same length as rows in the inDataFrameScaled. It is used to generate the final analysis, where a table of the percentage of observations for each individual and each cluster is created.}
}
\value{
A list with three components:
\describe{
    \item{clusterVector}{A vector with the same length as number of rows in the inDataFrameScaled, where the cluster identity of each observation is noted.}
    \item{penalizedClusterCenters}{A matrix containing information about which variables and clusters that have been penalizedd out, shown by zero, and for the non-penalized variables and clusters, where the cluster center is located. This is done with one row per cluster.}
    \item{clusterPercentagesForAllIds}{A matrix showing the percentage of observations for each id in each cluster.}
}
}
\description{
This function is the core user function of the Depeche package. It clusters the data with a penalized version of K-means.
}
\examples{
#Generate a default size dataframe with bimodally distributed data
x <- generateFlowCytometryData(samplings=2)

#Scale this datamframe
x_scaled <- quantileScale(x[,2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Run the pKMOptim function to get good starting points
x_optim <- pKMOptim(x_scaled, iterations=50, bootstrapObservations=1000)

#Then run the actual function
x_pKM <- pKMRun(x_scaled, regVec=x_optim[[1]][["optimalRegularizationValue"]], withOrWithoutZeroClust=x_optim[[1]][["withOrWithoutZeroClust"]], iterations=1, ids=x[,1])

#And finally look at your great result
str(x_pKM)
}
\seealso{
\code{\link{pKMOptim}}, \code{\link{pKMPredict}}
}
