% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData}
\alias{testData}
\title{Test dataset}
\format{An object of class \code{"data.frame"};}
\usage{
data(testData)
}
\description{
This dataset is a 14 color pre-compensated, transformed flow cytometry dataset focusing 
on cytotoxic lymphocytes, where dead cells have been removed. To make examples very 
obvious, differences have been artificially exaggregated in the data. The dataset is 
produced by J. Theorell.
}
\keyword{datasets}
