% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDataDepeche.R
\docType{data}
\name{testDataDepeche}
\alias{testDataDepeche}
\title{Test dataset}
\format{An object of class \code{"list"};}
\usage{
data(testDataDepeche)
}
\description{
This is a depeche clustering of the testData dataset. It has  
been generated with the Rtsne.multicore package
}
\details{
produced by J. Theorell.
}
\keyword{datasets}
