% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDataSNE.R
\docType{data}
\name{testDataSNE}
\alias{testDataSNE}
\title{SNE of the testData set}
\format{An object of class \code{"list"};}
\usage{
data(testDataSNE)
}
\description{
This is a t-distributed stochastic neighbor embedding of the testData dataset. It has  
been generated with the Rtsne.multicore package.
}
\references{
Jesse H. Krijthe (2015). Rtsne: T-Distributed Stochastic Neighbor 
Embedding using a Barnes-Hut Implementation
(\href{https://github.com/RGLab/Rtsne.multicore/issues/5}{GitHub})
}
\keyword{#'}
\keyword{datasets}
