% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateData.R
\name{truncateData}
\alias{truncateData}
\title{Truncate the extreme ends of numeric data}
\usage{
truncateData(x, control, lowQuantile = 1e-04, highQuantile = 0.9999)
}
\arguments{
\item{x}{A numeric/integer vector or dataframe.}

\item{control}{A numeric/integer vector or dataframe of values that could be used to define the range. If no control data is present, the function defaults to using x as control data. NB! The control file needs to be the same class, i.e vector or dataframe, as the x, and in the case of a dataframe it needs to have the same number of columns.}

\item{lowQuantile}{The lower border below which the values are treated as outliers and will be outside of the defined quantile range.}

\item{highQuantile}{The higher border above which the values are treated as outliers and will be outside of the defined quantile range.}
}
\value{
A vector or dataframe with the same size but where the extreme values have been substituted with the less extreme values.
}
\description{
This function reduces the most extremely low and high observations to their less extreme border. Currently, the standard quantile cutoff is a hundredth of a percent. It is done to decrease these observations influence on color scaling for vizualisations. The function accepts vectors and dataframes.
}
\examples{
#Generate a random vector
x <- rnorm(1000, 55, 10)

#The maximum and minimum values are
max(x)
min(x)

#Here, the function is used to truncate the two most extreme percent of the observations.
y <- truncateData(x, lowQuantile=0.01, highQuantile=0.99)

#The maximum and minimum values are less extreme:
max(y)
min(y)

#Do the same but with a dataframe
x_df <- data.frame(cbind(rnorm(1000, 55, 10), rnorm(1000, 2, 90), rnorm(1000, 430, 200)))
summary(x_df)

#Run the function
y_df <- truncateData(x_df, lowQuantile=0.01, highQuantile=0.99)

#And the most extreme values have been reduced
summary(y_df)

}
\seealso{
\code{\link{quantileScale}}
}
