% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dColorPlot.R
\name{dColorPlot}
\alias{dColorPlot}
\title{Display third variable as color on a 2D plot}
\usage{
dColorPlot(colorData, controlData, xYData, names = "default",
  densContour = TRUE, addLegend = FALSE, idsVector,
  drawColorPalette = FALSE, title = FALSE, createDirectory = TRUE,
  directoryName = "Variables displayed as color on SNE field",
  truncate = TRUE, bandColor = "black",
  dotSize = 500/sqrt(nrow(xYData)), multiCore = "default",
  createPlot = TRUE)
}
\arguments{
\item{colorData}{A vector, matrix or dataframe of numeric observations that will be displayed as color on the plot.}

\item{controlData}{Optional. A numeric/integer vector or dataframe of values that could be used to define the range of the colorData. If no control data is present, the function defaults to using the colorData as control data.}

\item{xYData}{These variables create the field on which the colorData will be displayed. It needs to be a matrix or dataframe with two columns and the same number of rows as the colorData object.}

\item{names}{The name(s) for the plots. The default alternative, 'default' returns the column names of the colorData object in the case this is a dataframe and otherwise returns the somewhat generic name 'testVariable'. It can be substituted with a string (in the case colorData is a vector) or vector of strings, as long as it has the same length as the number of columns in colorData.}

\item{densContour}{If density contours should be created for the plot(s) or not. Defaults to TRUE. a}

\item{addLegend}{If this is set to true, a separate legend plot is produced. This is most useful when the color data contains specific info about separate ids, such as clusters. Default is FALSE.}

\item{idsVector}{If a legend is added, this argument controls the naming in the legend.}

\item{drawColorPalette}{If a separate plot with the color palette used for the plots should be printed and saved.}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved a jpeg, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{truncate}{If truncation of the most extreme values should be performed for the visualizations. Three possible values: TRUE, FALSE, and a vector with two values indicating the low and high threshold quantiles for truncation.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}

\item{multiCore}{If the algorithm should be performed on multiple cores. This increases the speed if the dataset is medium-large (>100000 rows) and has at least 5 columns. Default is TRUE when the rows exceed 100000 rows and FALSE otherwise.}

\item{createPlot}{For testing purposes. Defaults to TRUE. If FALSE, no plots are generated.}
}
\value{
Plots showing the colorData displayed as color on the field created by xYData.
}
\description{
Function to overlay one variable for a set of observations on a field created by two other variables known for the same observations. The plot is constructed primarily for displaying variables on 2D-stochastic neighbour embedding fields, but can be used for any sets of (two or) three variables known for the same observations. As the number of datapoints is often very high, the files would, if saved as pdf of another vector based file type become extremely big. For this reason, the plots are saved as jpeg and no axes or anything alike are added, to simplify usage in publications.
}
\examples{
# Load some data
data(testData)

# Run Barnes Hut tSNE on this. For more rapid example execution, a pre-run
# SNE is inluded
# library(Rtsne)
# testDataSNE <- Rtsne(testData[,2:15], pca=FALSE)
data(testDataSNE)

# Run the function for two of the variables
dColorPlot(colorData = testData[2:3], xYData = testDataSNE$Y,
drawColorPalette = TRUE)

# Create a color vector and display it on the SNE field. For this purpose,
# four individual donors are extracted from the test dataset
testDataSubset <- rbind(testData[1:2000, ], testData[95001:97000, ])
testDataSNESubset <- rbind(testDataSNE$Y[1:2000, ], testDataSNE$Y[95001:97000, ])
testColor <- dColorVector(testDataSubset$ids, colorScale = 'plasma')
dColorPlot(
  colorData = testColor, xYData = testDataSNESubset,
  names = 'separate samplings', addLegend = TRUE, idsVector = testDataSubset$ids
)
}
\seealso{
\code{\link{dDensityPlot}}, \code{\link{dResidualPlot}}, \code{\link{dWilcox}}, \code{\link{dColorVector}}
}
