% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depecheDensity.R
\name{depecheDensity}
\alias{depecheDensity}
\title{Display density on 2D plot}
\usage{
depecheDensity(xYData, color = c("blue", "rainbowCols", "a colorVector"),
  commonName, plotEachIdSeparately = FALSE, idsVector, densContour,
  title = FALSE, createDirectory = TRUE,
  directoryName = paste("Density plots for ", commonName, "s", sep = ""),
  scalingControl, bandColor = "black", dotSize = 400/sqrt(nrow(xYData)))
}
\arguments{
\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation.}

\item{color}{This gives the specific color of the plot. It has three potential return values:
\describe{
    \item{A specific color, e.g. "red" or "#FF0000"}{This will be the color in the densest part of the plot(s)}
    \item{"rainbowCols"}{Creates a specific set of colors similar to the rainbow.}
    \item{A vector of colors, the same length as nrow(xYData)}{Here, a separate density estimate will be performed for each color subcompartment of the xYData and these will be plotted together.}
}}

\item{commonName}{A name that is common to all density plots created. It can be the groups name, e.g. "Malaria patients" or "Clusters". If only one plot is created, the name is still taken from here.}

\item{plotEachIdSeparately}{Separates the xYData into subframes specified in the idsVector and plots all ids together but with separate colors. Especially useful when all clusters in a simultaneous analysis should be shown. Colors are inherited from color. Defaults to FALSE.}

\item{idsVector}{Needed in two situations:
\describe{
    \item{When "plotEachIdSeparately"=TRUE}{Provides information about which rows that belong to which id and the names for the individual plots.}
    \item{When "color" is a vector of colors}{The ids are used to create the legend.}
}}

\item{densContour}{An object to create the density contours for the plot. If not present, it will be generated with the xYData. Useful when only a subfraction of a dataset is plotted, and a superimposition of the distribution of the whole dataset is of interest.}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{scalingControl}{A dataframe with two columns. This argument is only necessary if the xYData should be scaled to another range than its internal maximum and minimum values.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}
}
\value{
Plots showing the densities of the specific xYData (subset) displayed as color on the field created by the same xYData (subset).
}
\description{
Function to show density for a set of observations on a field created by two variables. The plot is constructed primarily for displaying density of 2D-stochastic neighbour embedding fields, but can be used for any sets of two known for the same observations. As the number of datapoints is often very high, the files would, if saved as pdf of another vector based file type become big. For this reason, the plots are saved as jpeg and no axes or anything alike are added, to simplify usage in publications.
}
\examples{
#Generate a dataframe with bimodally distributed data and a few separate subsamplings
x <- generateFlowCytometryData(samplings=5, observations=2000)

#Scale the data (not actually necessary in this artificial 
#example due to the nature of the generated data)
x_scaled <- quantileScale(x=x[2:ncol(x)])

#Run Barnes Hut tSNE on this. 
library(Rtsne.multicore)
xSNE <- Rtsne.multicore(x_scaled, pca=FALSE)

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Plot all ids together and use a fixed color
depecheDensity(xYData=as.data.frame(xSNE$Y), commonName="All_samplings", 
color="blue", createDirectory=FALSE)

#Now plot each id separately using a predefined colorscale separating each cluster
xColor <- colorVector(x[,1], colorScale="plasma")
depecheDensity(xYData=as.data.frame(xSNE$Y), color=xColor, plotEachIdSeparately=TRUE, 
idsVector=x[,1], commonName="sampling")

#Now all clusters are plotted together using the same predefined colorscale
depecheDensity(xYData=as.data.frame(xSNE$Y), color=xColor, idsVector=x[,1],
commonName="all samplings")

}
\seealso{
\code{\link{depecheColor}}, \code{\link{depecheResidual}}, \code{\link{depecheWilcox}}
}
