% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dResidualPlot.R
\name{dResidualPlot}
\alias{dResidualPlot}
\title{Showing the residuals when subtracting the values from one group from another on a sne plot}
\usage{
dResidualPlot(xYData, groupVector, clusterVector, densContour = TRUE,
  name = "dResidualPlot", groupName1 = unique(groupVector)[1],
  groupName2 = unique(groupVector)[2], title = FALSE, maxAbsPlottingValues,
  bandColor = "black", createDirectory = FALSE,
  directoryName = "dResidualPlot", dotSize = 400/sqrt(nrow(xYData)))
}
\arguments{
\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation.}

\item{groupVector}{Vector with the same length as xYData containing information about the group identity of each observation.}

\item{clusterVector}{Vector with the same length as xYData containing information about the cluster identity of each observation.}

\item{densContour}{An object to create the density contours for the plot. Three possible values: 
\describe{
              \item{densContour}{A densContour object generated previously with dContours}
              \item{TRUE}{a densContour object will be generated internally}
              \item{FALSE}{No density contours will be displayed.}
             }}

\item{name}{The main name for the graph and the analysis.}

\item{groupName1}{The name for the first group}

\item{groupName2}{The name for the second group}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{maxAbsPlottingValues}{If multiple plots should be compared, it might be useful to define a similar color scale for all plots, so that the same color always means the same value. Such a value can be added here. It defaults to the maximum Wilcoxon statistic that is generated in the analysis.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}
}
\value{
A sne based plot showing which events that belong to a cluster dominated by the first or the second group.
}
\description{
This function is used to compare groups of individuals from whom comparable cytometry or other complex data has been generated where the number of individuals does not allow any statistical comparisons.
}
\examples{
#Generate a dataframe with bimodally distributed data and 2 subsamplings.
x <- generateBimodalData(samplings=2, ncols=7)

#Scale the data 
x_scaled <- dScale(x=x[2:ncol(x)])

#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Optimize and run the clustering function.
xOptAndClustObject <- dOptAndClust(x_scaled,ids=x[,1])
xClustObject <- xOptAndClustObject[[2]]

#Run Barnes Hut tSNE on this. 
library(Rtsne.multicore)
xSNE <- Rtsne.multicore(x_scaled, pca=FALSE)

#And finally run the function
dResidualPlot(xYData=as.data.frame(xSNE$Y), groupVector=x[,1], clusterVector=xClustObject$clusterVector)
}
\seealso{
\code{\link{dColorPlot}}, \code{\link{dDensityPlot}}, \code{\link{dWilcoxPlot}}
}
