% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSplsda.R
\name{dSplsda}
\alias{dSplsda}
\title{Sparse partial least squares discriminant analysis with paired and unpaired data}
\usage{
dSplsda(xYData, idsVector, groupVector, clusterVector, pairingVector,
  displayVector, densContour = TRUE, name = "dSplsda",
  groupName1 = unique(groupVector)[1], groupName2 = unique(groupVector)[2],
  title = FALSE, maxAbsPlottingValues, createDirectory = FALSE,
  directoryName = "dSplsda", bandColor = "black",
  dotSize = 400/sqrt(nrow(xYData)))
}
\arguments{
\item{xYData}{A dataframe with two columns. Each row contains information about the x and y positition in the field for that observation.}

\item{idsVector}{Vector with the same length as xYData containing information about the id of each observation.}

\item{groupVector}{Vector with the same length as xYData containing information about the group identity of each observation.}

\item{clusterVector}{Vector with the same length as xYData containing information about the cluster identity of each observation.}

\item{pairingVector}{If this vector is present, a multilevel spls-da will be performed, that considers the within-donor variation between different stimuli. Defaults to NULL.}

\item{displayVector}{Optionally, if the dataset is very large and the SNE calculation hence becomes impossible to perform for the full dataset, this vector can be included. It should contain the set of rows from the data used for statistics, that has been used to generate the xYData.}

\item{densContour}{An object to create the density contours for the plot. If not present, it will be generated with the xYData. Useful when only a subfraction of a dataset is plotted, and a superimposition of the distribution of the whole dataset is of interest.}

\item{name}{The main name for the graph and the analysis.}

\item{groupName1}{The name for the first group}

\item{groupName2}{The name for the second group}

\item{title}{If there should be a title displayed on the plotting field. As the plotting field is saved as a png, this title cannot be removed as an object afterwards, as it is saved as coloured pixels. To simplify usage for publication, the default is FALSE, as the files are still named, eventhough no title appears on the plot.}

\item{maxAbsPlottingValues}{If multiple plots should be compared, it might be useful to define a similar color scale for all plots, so that the same color always means the same statistical value. Such a value can be added here. It defaults to the maximum statistic that is generated in the analysis.}

\item{createDirectory}{If a directory (i.e. folder) should be created. Defaults to TRUE.}

\item{directoryName}{The name of the created directory, if it should be created.}

\item{bandColor}{The color of the contour bands. Defaults to black.}

\item{dotSize}{Simply the size of the dots. The default makes the dots smaller the more observations that are included.}

\item{densContour}{An object to create the density contours for the plot. Three possible values: 
\describe{
              \item{densContour}{A densContour object generated previously with dContours}
              \item{TRUE}{a densContour object will be generated internally}
              \item{FALSE}{No density contours will be displayed.}
             }}
}
\value{
This function returns the full result of the sPLS-DA. It also returns a sne based plot showing which events that belong to a cluster dominated by the first or the second group defined by the sparse partial least squares loadings of the clusters.
}
\description{
This function is used to compare groups of individuals from whom comparable cytometry or other complex data has been generated. It is superior to just running a Wilcoxon analysis in that it does not consider each cluster individually, but instead uses a sparse partial least squares discriminant analysis to first identify which vector thourgh the multidimensional data cloud, created by the cluster-donor matrix, that optimally separates the groups, and as it is a sparse algorithm, applies a penalty to exclude the clusters that are orthogonal, or almost orthogonal to the discriminant vector, i.e. that do not contribute to separating the groups.
}
\examples{
#Generate a dataframe with bimodally distributed data and 20 subsamplings.
xindividuals <- generateBimodalData(samplings=40, dataCols=7, observations=500)

#Now add three columns that will separate the first ten from 
#the second ten individuals and merge the datasets
xgroups <- generateBimodalData(samplings=2, dataCols=3, observations=10000)
colnames(xgroups)[2:4] <- c("X8", "X9", "X10")
x <- cbind(xindividuals[,1], xgroups[,1], 
xindividuals[,2:ncol(xindividuals)], xgroups[,2:ncol(xgroups)])

colnames(x)[1:2] <- c("ids", "group")

#Scale the data
x_scaled <- dScale(x[3:ncol(x)])
#Set a reasonable working directory, e.g.
setwd("~/Desktop")

#Optimize and run the clustering function.
xClustObject <- dClust(x_scaled, sampleSizes=100, selectionSampleSize=1000, maxIter=20)
clusterVector <- xClustObject[[1]]

#Run Barnes Hut tSNE on this. 
library(Rtsne.multicore)
xSNE <- Rtsne.multicore(x_scaled, pca=FALSE)

#Run the function. This time without pairing.
sPLSDAObject <- dSplsda(xYData=as.data.frame(xSNE$Y), idsVector=x$ids, groupVector=x$group, clusterVector=clusterVector)

#Here, pairing is used
#First, an artificial pairing vector, making the first donor amongst the first ten connected to the first donor among the second ten.
pairingVector <- c(rep(1:20, each=500), rep(1:20, each=500))

#Then the actual multilevel sPLS-DA is run. 
sPLSDAObject <- dSplsda(xYData=as.data.frame(xSNE$Y), idsVector=x$ids, groupVector=x$group, clusterVector=clusterVector, pairingVector=pairingVector, name="d_sPLSDAPlot_paired", groupName1="Stimulation 1", groupName2="Stimulation 2")
}
\seealso{
\code{\link{dColorPlot}}, \code{\link{dDensityPlot}}, \code{\link{dResidualPlot}}
}
